+++ CHANGELOG.rst
@@ -15,7 +15,11 @@ Changelog
 
 Added
 ^^^^^
+- Add Case-When support. (#943)
 - Add `Rand`/`Random` function in contrib. (#944)
+Changed
+^^^^^^^
+- Remove `Function`,`Aggregate` from 'functions.py' to 'expressions.py'. (#943)
 
 0.17.8
 ------